requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic>=2.3
pdfplumber
pytesseract
pillow
python-dotenv
openai>=1.30.0
python-dateutil
regex
starlette
reportlab
